+++ Contributing.md
@@ -1,4 +1,4 @@
-# Как работать с Git на проекте
+пше# Как работать с Git на проекте
 
 Поддерживайте ваш репозиторий обновлённым. Когда наставник принимает ваш пулреквест, он попадает в репозиторий Академии, но не в ваш форк.
